Move ClassUtils location into KotlinDetector
