+++ core/src/main/java/feign/kotlinSupport/KotlinDetector.java
@@ -35,4 +35,31 @@ public abstract class KotlinDetector {
 	public static boolean isUnitType(Type type) {
 		return supportsKotlin && Unit.class == type;
 	}
+
+	private static class ClassUtils {
+		public static Class<?> tryGetForName(String name, ClassLoader classLoader) {
+			try {
+				return Class.forName(name, false, classLoader);
+			}
+			catch (ClassNotFoundException | NoClassDefFoundError ex) {
+				return null;
+			}
+		}
+
+		public static boolean isPresent(String className, ClassLoader classLoader) {
+			try {
+				return tryGetForName(className, classLoader) != null;
+			}
+			catch (IllegalAccessError err) {
+				String message = String.format(
+						"Readability mismatch in inheritance hierarchy of class [%s]: %s",
+						className, err.getMessage()
+				);
+				throw new IllegalStateException(message, err);
+			}
+			catch (Throwable ex) {
+				return false;
+			}
+		}
+	}
 }
